# src/main/java/com/example/bankcards/dto/CardDto.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.enums.CardStatus;
import java.math.BigDecimal;
import java.time.LocalDate;

/**
 * DTO с информацией о банковской карте.
 */
public record CardDto(
        Long id,
        String maskedNumber,
        LocalDate expirationDate,
        CardStatus status,
        BigDecimal balance
) {
}