# src/main/java/com/example/bankcards/dto/UpdateCardRequestDto.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.enums.CardStatus;
import jakarta.validation.constraints.Future;
import jakarta.validation.constraints.NotNull;
import java.time.LocalDate;

/**
 * Запрос на обновление информации о банковской карте.
 */
public record UpdateCardRequestDto(
        @NotNull(message = "Expiration date is required")
        @Future(message = "Expiration date must be in the future")
        LocalDate expirationDate,
        CardStatus status
) {
}